+++ pyrava/__init__.py
@@ -54,7 +54,7 @@ from .packet import (
 )
 from .transport import HttpTransport, Transport
 
-__version__ = "0.1.5"
+__version__ = "0.2.0"
 
 __all__ = [
     "__version__",